Fix Orientation
Make the orientation of LNNotificationBannerWindow changed base on
statusBarOrientation
Sometimes, I hope the UIViewController force landscape. When trigger
the LNNotification, it maybe portrait or landscape base on device. The
LNNotification's Orientation is different with the UIViewController. If
the LNNotification's Orientation is the same with the statusbar, the
problem can be solved.