Fix Xcode 7 warning
Current build generated the warning: “Method override for the
designated initializer of the superclass ‘init’ not found”.

According to Apple Docs: “If a class provides one or more designated
initializers, it must implement all of the designated initializers of
its superclass.”

I have added an override for the superclass designated initializer
(init) which calls your designated initializer with nil parameters. If
you prefer to make init unavailable you could add to there header
instead:

- (instancetype)init NS_UNAVAILABLE;

It’s a judgement call which way to go.